// apps/web/src/lib/api/generated/model/paymentInMethod.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export type PaymentInMethod = typeof PaymentInMethod[keyof typeof PaymentInMethod];


export const PaymentInMethod = {
  mpesa: 'mpesa',
  bank: 'bank',
  card: 'card',
  cheque: 'cheque',
  cash: 'cash',
  other: 'other',
} as const;