// apps/web/src/lib/api/generated/model/paymentOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BrokerOS API
 * OpenAPI spec version: 1.0.0
 */

export interface PaymentOut {
  /** @pattern ^-?\d{1,16}(\.\d{1,4})?$ */
  amount: string;
  created_at: string;
  created_by: string | null;
  currency: string;
  id: string;
  method: string;
  paid_on: string;
  paid_to: string;
  reference: string | null;
  remittance_reference: string | null;
  remitted_on: string | null;
  void_reason: string | null;
  voided_at: string | null;
}